safari-534.53-branch/LayoutTests/platform/qt/fast/events/click-focus-anchor-expected.txt

This test ensures that we can click to focus an a element. Click on the element below.

The expected result is platform specific. GTK and QT allows all links to be mouse focusable.

Result


a1 received focus (focusable)
a3 received focus (focusable on GTK and QT)
a4 received focus (focusable)
a5 received focus (focusable)
